+++ app/configurations/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from .models import XDSConfiguration, XDSUIConfiguration
+from .models import (CourseInformationMapping, XDSConfiguration,
+                     XDSUIConfiguration)
 
 
 # Register your models here.
@@ -16,3 +17,12 @@ class XDSUIConfigurationAdmin(admin.ModelAdmin):
                     'created', 'modified',)
     fields = [('search_results_per_page', 'xds_configuration',
                'course_img_fallback')]
+
+
+@admin.register(CourseInformationMapping)
+class CourseInformationMappingAdmin(admin.ModelAdmin):
+    list_display = ('course_title', 'course_description',
+                    'course_url', 'xds_ui_configuration')
+
+    fields = ['course_title', 'course_description',
+              'course_url', 'xds_ui_configuration']
