+++ app/configurations/tests/test_views_unit.py
@@ -5,8 +5,8 @@ from django.test import tag
 from django.urls import reverse
 from rest_framework import status
 
-from configurations.models import XDSUIConfiguration
-from configurations.models import XDSConfiguration
+from configurations.models import (XDSConfiguration, XDSUIConfiguration,
+                                   CourseInformationMapping)
 
 from .test_setup import TestSetUp
 
@@ -54,3 +54,25 @@ class ModelTests(TestSetUp):
         uiConfig = XDSUIConfiguration(xds_configuration=config)
 
         self.assertEqual(uiConfig.search_results_per_page, 10)
+
+    def test_create_courseInformationMapping(self):
+        """Tests the creation of a course information object"""
+
+        config = XDSConfiguration(target_xis_metadata_api="test")
+        uiConfig = XDSUIConfiguration(xds_configuration=config)
+
+        # course mappings
+        course_title = 'Course.TestTitle'
+        course_description = 'Course.TestDescription'
+        course_url = 'Course.TestUrl'
+
+        courseInformation = CourseInformationMapping(
+            xds_ui_configuration=uiConfig,
+            course_title=course_title,
+            course_description=course_description,
+            course_url=course_url)
+
+        self.assertEqual(courseInformation.course_title, course_title)
+        self.assertEqual(courseInformation.course_description,
+                         course_description)
+        self.assertEqual(courseInformation.course_url, course_url)
