+++ app/core/models.py
@@ -143,44 +143,6 @@ class CourseSpotlight(TimeStampedModel):
         return f'{self.id}'
 
 
-class CourseInformationMapping(TimeStampedModel):
-    """ Model to map course information"""
-
-    course_title = models.CharField(max_length=200,
-                                    help_text="Enter the title of the course"
-                                              "found in the elasticsearch")
-    course_description = models.CharField(max_length=200,
-                                          help_text="Enter the description of"
-                                                    " the course found in the"
-                                                    " elasticsearch")
-    course_url = models.CharField(max_length=200,
-                                  help_text="Enter the url of the course found"
-                                            " in the elasticsearch")
-
-    xds_ui_configuration = models \
-        .OneToOneField(XDSUIConfiguration,
-                       on_delete=models.CASCADE,
-                       related_name='course_information')
-
-    def get_absolute_url(self):
-        """ URL for displaying individual model records."""
-        return reverse('course-information', args=[str(self.id)])
-
-    def __str__(self):
-        """String for representing the Model object."""
-        return f'{self.id}'
-
-    def save(self, *args, **kwargs):
-        num_active_mappings = CourseInformationMapping.objects.filter().count()
-
-        # if there is more than one
-        if num_active_mappings > 1:
-            raise ValidationError(
-                'Max of 1 active highlight fields has been reached.')
-
-        return super(CourseInformationMapping, self).save(*args, **kwargs)
-
-
 class Experience(models.Model):
     """Model to store experience instances for interest lists"""
 
@@ -215,131 +177,3 @@ class SavedFilter(TimeStampedModel):
                             help_text="Enter the name of the filter")
     query = models.CharField(max_length=200,
                              help_text="queryString for the filter")
-
-
-# class PermissionsChecker(DjangoModelPermissions):
-#     """
-#     Class to define the method for checking permissions for the XDS API
-#     """
-#     perms_map = {
-#         'GET': ['%(app_label)s.view_%(model_name)s'],
-#         'OPTIONS': ['%(app_label)s.view_%(model_name)s'],
-#         'POST': ['%(app_label)s.add_%(model_name)s'],
-#         'PUT': ['%(app_label)s.change_%(model_name)s'],
-#         'PATCH': ['%(app_label)s.change_%(model_name)s'],
-#         'DELETE': ['%(app_label)s.delete_%(model_name)s'],
-#     }
-
-#     def has_permission(self, request, view):
-#         # Workaround to ensure DjangoModelPermissions are not applied
-#         # to the root view when using DefaultRouter.
-#         if getattr(view, '_ignore_model_permissions', False):
-#             return True
-
-#         # if current request is in OPEN_ENDPOINTS doesn't check permissions,
-#         # returns true
-#         if request.path_info in getattr(settings, 'OPEN_ENDPOINTS', []):
-#             return True
-
-#         # checks if there is a logged in user
-#         if not request.user or (
-#                 not request.user.is_authenticated and
-#                 self.authenticated_users_only):
-#             return False
-
-#         try:
-#             # tries to get app and model names from view
-#             model_meta = self._queryset(view).model._meta
-
-#         except Exception:
-#             # if unable, generates app and model names
-#             def model_meta():
-#                 return None
-
-#             model_meta.app_label = "core"
-#             model_meta.model_name = \
-#                 view.get_view_name().lower().replace(' ', '')
-
-#         # determines permission required to access this endpoint
-#         perms = self.get_required_permissions(request.method, model_meta)
-
-#         # checks if the user has the required permission
-#         return request.user.has_perms(perms)
-
-#     def get_required_permissions(self, method, model_meta):
-#         """
-#         Given a model and an HTTP method, return the list of permission
-#         codes that the user is required to have.
-#         """
-#         kwargs = {
-#             'app_label': model_meta.app_label,
-#             'model_name': model_meta.model_name
-#         }
-
-#         if method not in self.perms_map:
-#             raise exceptions.MethodNotAllowed(method)
-
-#         return [perm % kwargs for perm in self.perms_map[method]]
-
-
-# class NumberValidator(object):
-#     def validate(self, password, user=None):
-#         if not re.findall('\\d', password):
-#             raise ValidationError(
-#                 _("The password must contain at least 1 digit, 0-9."),
-#                 code='password_no_number',
-#             )
-
-#     def get_help_text(self):
-#         return _(
-#             "Your password must contain at least 1 digit, 0-9."
-#         )
-
-
-# class UppercaseValidator(object):
-#     def validate(self, password, user=None):
-#         if not re.findall('[A-Z]', password):
-#             raise ValidationError(
-#                 _(
-#                     "The password must contain at least 1 uppercase letter, "
-#                     "A-Z."),
-#                 code='password_no_upper',
-#             )
-
-#     def get_help_text(self):
-#         return _(
-#             "Your password must contain at least 1 uppercase letter, A-Z."
-#         )
-
-
-# class LowercaseValidator(object):
-#     def validate(self, password, user=None):
-#         if not re.findall('[a-z]', password):
-#             raise ValidationError(
-#                 _(
-#                     "The password must contain at least 1 lowercase letter, "
-#                     "a-z."),
-#                 code='password_no_lower',
-#             )
-
-#     def get_help_text(self):
-#         return _(
-#             "Your password must contain at least 1 lowercase letter, a-z."
-#         )
-
-
-# class SymbolValidator(object):
-#     def validate(self, password, user=None):
-#         if not re.findall('[()[\\]{}|\\\\`~!@#$%^&*_\\-+=;:\'",<>./?]',
-#                           password):
-#             raise ValidationError(
-#                 _("The password must contain at least 1 symbol: " +
-#                   "()[]{}|\\`~!@#$%^&*_-+=;:'\",<>./?"),
-#                 code='password_no_symbol',
-#             )
-
-#     def get_help_text(self):
-#         return _(
-#             "Your password must contain at least 1 symbol: " +
-#             "()[]{}|\\`~!@#$%^&*_-+=;:'\",<>./?"
-#         )
